requirements.txt
asttokens==2.4.1
azure-core==1.30.1
azure-storage-blob==12.20.0
azure-storage-file-datalake==12.15.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
et-xmlfile==1.1.0
executing==2.0.1
fonttools==4.53.0
fpdf2==2.7.9
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
isodate==0.6.1
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
msal==1.28.0
nest-asyncio==1.6.0
numpy==2.0.0
Office365-REST-Python-Client==2.5.11      
openpyxl==3.1.5
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyodbc==5.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==26.0.3
reportlab==4.2.2
requests==2.31.0
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
xlrd==2.0.1
xlwings==0.31.10